refactor: jparepository 구조 수정
