+++ src/main/java/com/catcher/core/service/ScheduleService.java
@@ -1,18 +1,16 @@
 package com.catcher.core.service;
 
 import com.catcher.core.domain.entity.Schedule;
-import com.catcher.datasource.ScheduleRepository;
+import com.catcher.datasource.schedule.ScheduleJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.time.ZonedDateTime;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class ScheduleService {
-
-    private final ScheduleRepository scheduleRepository;
+    private final ScheduleJpaRepository scheduleRepository;
 
     public List<Schedule> getScheduleByKeywordAndFilter(
             final String keyword
@@ -28,7 +26,5 @@ public class ScheduleService {
 //                endAt
         );
         return schedules;
-
     }
-
 }
