--- src/main/java/com/catcher/datasource/ScheduleRepository.java
@@ -1,11 +1,0 @@
-package com.catcher.datasource;
-
-import com.catcher.core.domain.entity.Schedule;
-import com.catcher.datasource.custom.ScheduleRepositoryCustom;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ScheduleRepository extends JpaRepository<Schedule, Long>, ScheduleRepositoryCustom {
-
-}
