+++ src/main/java/com/catcher/datasource/schedule/ScheduleRepositoryImpl.java
@@ -1,15 +1,15 @@
-package com.catcher.datasource.custom;
+package com.catcher.datasource.schedule;
 
+import com.catcher.core.db.ScheduleRepository;
 import com.catcher.core.domain.entity.QSchedule;
 import com.catcher.core.domain.entity.Schedule;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
 
-import java.time.ZonedDateTime;
 import java.util.List;
 
-public class ScheduleRepositoryImpl extends QuerydslRepositorySupport implements ScheduleRepositoryCustom {
+public class ScheduleRepositoryImpl extends QuerydslRepositorySupport implements ScheduleRepository {
 
     private final JPAQueryFactory jpaQueryFactory;
     private static final QSchedule SCHEDULE = QSchedule.schedule;
