+++ src/main/java/com/catcher/resource/ScheduleController.java
@@ -4,10 +4,7 @@ import com.catcher.common.response.CommonResponse;
 import com.catcher.core.domain.entity.Schedule;
 import com.catcher.core.service.ScheduleService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 @RestController
